requirements.txt
opencv-python
numpy
matplotlib
PyYAML
scikit-image
shapely